=== New File: Player_Manager.cs
This file will contain the methods that are intended to be used by players and AI when drawing resource cards, route cards, and when placing track down.  The GameManager will make calls to these methods passing it what player has the current turn.

I am expecting to extend this class to the AI_Manager as the AI_Manager will need to utilize these methods along with its own unique ones.

====
GameManager.cs

I edited what was there for the game manager.  The structure done by Randy was correct. Just answered some questions he had by adding a comment at the top of the document detailing intentions.
Also replaced code with their correct implementations for use with the new Player_Manager file.
If there are any questiosn regarding these changes please let me know.

====
====
Card.cs

I created a Hand<T> class.  This is to be used to store exactly what cards each player has.  The T will be for use with resource cards and route cards.
I still need to add functionality of the Remove method within the Hand<T> class.

Also added a Route card deck, but still need to add the code for the actual creation of its cards.
====
====
PlayerList.cs

I made a few changes to this script.  I changed the public class PlayerInfo to be a struct as it appears to fit the description of a struct more than a class.  Also changed the values that stored an int count of the total resource cards and route cards each player hand to be a Hand<T> object.

Also, made a simple readability change to the way labels are setup as well as data was initialized in the struct.  Instead of using a boolean value "isBot" I changed it to "isHuman" to fit more with the boolean value we store within the PlayerInfo to represent that.
====